Move UnresolvedAttributes to reference module
